backend/pyproject.toml
[tool.poetry]
name = "online-store"
version = "0.1.0"
description = ""
authors = ["kooznitsa"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
gunicorn = "^21.2.0"
uvicorn = "^0.28.1"
aiohttp = "^3.9.3"
python-dotenv = "^1.0.1"
pillow = "^10.2.0"
psycopg2-binary = "^2.9.9"
celery = "^5.3.6"
flower = "^2.0.1"
stripe = "^8.7.0"
django-cors-headers = "^4.3.1"
xhtml2pdf = "^0.2.15"
redis = "^5.0.3"
celery-amqp-backend = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
